Have SourceBufferPrivateAVJObjC use WorkQueue instead of dispatch_async https://bugs.webkit.org/show_bug.cgi?id=230881 rdar://83429276
This change is twofold:
1- dispatch_async takes an objective-C block which doesn't support C++ move
semantics. As such the lambda would be copied including the captured variables.
When use with a non-thread safe refcounted object it could lead to a race
on the refcount which would result in a leak.
By using a WorkQueue instead, we avoid the above problem.
2- We want to unify in the media code how threads are created and task
dispatched: WorkQueue is a nice solution.

Reviewed by Chris Dumez.

No observable difference.

* platform/graphics/avfoundation/objc/SourceBufferPrivateAVFObjC.h:
* platform/graphics/avfoundation/objc/SourceBufferPrivateAVFObjC.mm:
(WebCore::SourceBufferPrivateAVFObjC::SourceBufferPrivateAVFObjC):
(WebCore::SourceBufferPrivateAVFObjC::append):
(WebCore::SourceBufferPrivateAVFObjC::resetParserState):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283685 268f45cc-cd09-0410-ab3c-d52691b4dbfc