[LFC][Integration] Optimize InlineContent::boxesForRect https://bugs.webkit.org/show_bug.cgi?id=231356
Reviewed by Alan Bujtas.

* layout/integration/LayoutIntegrationInlineContent.cpp:
(WebCore::LayoutIntegration::InlineContent::boxesForRect const):

Approximate the starting points from the rect using the average line height and then search for
the first and last line that paint fully outside the rect.

* layout/integration/LayoutIntegrationInlineContent.h:
* layout/integration/LayoutIntegrationInlineContentBuilder.cpp:
(WebCore::LayoutIntegration::InlineContentBuilder::createDisplayLines const):

Add a bit that tells if the content has lines with paint area exceeding those of the neighboring lines.
We can't use the optimization in this rare case.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283711 268f45cc-cd09-0410-ab3c-d52691b4dbfc