De-virtualize two overrides of GraphicsContext::drawImage https://bugs.webkit.org/show_bug.cgi?id=231399
Reviewed by Devin Rousso.

* platform/graphics/GraphicsContext.h:
(WebCore::GraphicsContext::drawImage):
* platform/graphics/NullGraphicsContext.h:
Devin correctly noted in https://bugs.webkit.org/show_bug.cgi?id=231385#c5
that it doesn't make sense for the two variants of drawImage() that
just call the version with more arguments to be virtual, and indeed, they
are not overridden. Devirtualize them, and remove the NullGraphicsContext overrides.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283762 268f45cc-cd09-0410-ab3c-d52691b4dbfc