[LibWPE] Do not destroy EGL backend if not created https://bugs.webkit.org/show_bug.cgi?id=224309
Reviewed by Carlos Garcia Campos.

about:gpu will create a PlatformDisplay to fill in GPU
information, but this doesn't call
LibWPEPlatformDisplay::initialize() as it doesn't need to, so the
backend member is not initialized. Make sure that the backend is
not a nullptr before attempting to destroy it when the display is
destroyed.

* platform/graphics/libwpe/PlatformDisplayLibWPE.cpp:
(WebCore::PlatformDisplayLibWPE::~PlatformDisplayLibWPE):
* platform/graphics/libwpe/PlatformDisplayLibWPE.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283801 268f45cc-cd09-0410-ab3c-d52691b4dbfc