Remove scrollbars explicitly when destroying render tree https://bugs.webkit.org/show_bug.cgi?id=229274
Patch by Rob Buis <rbuis@igalia.com> on 2021-10-09
Reviewed by Simon Fraser.

Source/WebCore:

Scrollbars in FrameViews that are hosted by RenderWidget need the RenderView
to exist because of RenderScrollbarPart. So when we are destroying the render tree
the RenderView will be destroyed too, so before that happens remove the scrollbars
and its RenderScrollbarParts.

Test: editing/inserting/insert-html-crash-02.html

* page/FrameView.cpp:
(WebCore::FrameView::willBeDestroyed):
* page/FrameView.h:
* platform/Widget.h:
(WebCore::Widget::willBeDestroyed):
* rendering/RenderWidget.cpp:
(WebCore::RenderWidget::willBeDestroyed):

LayoutTests:

* editing/inserting/insert-html-crash-02-expected.txt: Added.
* editing/inserting/insert-html-crash-02.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283868 268f45cc-cd09-0410-ab3c-d52691b4dbfc