[JSC][ARMv7] Make LLInt CSR save/restore order match the JITs https://bugs.webkit.org/show_bug.cgi?id=231517
Patch by Geza Lore <glore@igalia.com> on 2021-10-11
Reviewed by Saam Barati.

Since the new unlinked baseline JIT was introduced, the baseline JIT
uses the metadata table GPR the same way as the LLInt does, and relies
on it being in a saved CSR in emit_op_catch. This unearthed an issue
in the order in which LLInt on ARMv7 stores and restores CSRs when
trying to handle exceptions. This patch fixes the mismatch.

* llint/LowLevelInterpreter.asm:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@283914 268f45cc-cd09-0410-ab3c-d52691b4dbfc