Remove adjustForIOSCaretWhenScrolling() code https://bugs.webkit.org/show_bug.cgi?id=230454
Reviewed by Simon Fraser.

Source/WebCore:

The "adjustForIOSCaretWhenScrolling" code attempted to over-scroll horizontally at the start
or end of an overflow scroll in order to make space for the caret to show. However, it was
wrong in various ways. It always did this, even for scrollers with no visible caret. It was
web-detectable that the scrollX for a scroller could end up as -2. And removing the code
doesn't result in obviously incorrect behavior.

* editing/FrameSelection.cpp:
(WebCore::FrameSelection::revealSelection):
* rendering/RenderLayerScrollableArea.cpp:
(WebCore::RenderLayerScrollableArea::scrollTo):
(WebCore::RenderLayerScrollableArea::updateScrollInfoAfterLayout):
* rendering/RenderLayerScrollableArea.h:

Source/WebKitLegacy/mac:

The "adjustForIOSCaretWhenScrolling" code attempted to over-scroll horizontally at the start
or end of an overflow scroll in order to make space for the caret to show. However, it was
wrong in various ways. It always did this, even for scrollers with no visible caret. It was
web-detectable that the scrollX for a scroller could end up as -2. And removing the code
doesn't result in obviously incorrect behavior.

* DOM/DOMHTML.mm:
(-[DOMHTMLElement setScrollXOffset:scrollYOffset:adjustForIOSCaret:]):
* WebView/WebFrame.mm:
(-[WebFrame _scrollDOMRangeToVisible:]):
(-[WebFrame _scrollDOMRangeToVisible:withInset:]):

LayoutTests:

Rebase tests with the unified code path.

* platform/ios-wk2/editing/input/caret-at-the-edge-of-input-expected.txt:
* platform/ios-wk2/fast/layers/scroll-rect-to-visible-expected.txt:
* platform/ios/fast/forms/input-text-scroll-left-on-blur-expected.txt:
* platform/ios/fast/forms/textfield-outline-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284138 268f45cc-cd09-0410-ab3c-d52691b4dbfc