REGRESSION (r283858): Intense white hover state appears on playback controls on Netflix/YouTube https://bugs.webkit.org/show_bug.cgi?id=231719 rdar://84213404
Reviewed by Tim Horton.

Source/WebCore:

The playback controls on Netflix/YouTube are buttons that have a
non-'none' appearance value, but a used value of 'none' due to author
styles. In WebKit, and other engines, setting a background or text
color different from the UA style makes the used value 'none' and
drops native appearance.

The used value (effectiveAppearance) is set to none by comparing the
element's style to the UA style in RenderTheme::adjustStyle. However, when
style is adjusted during animation, the UA style is not passed in to the
method (see r273003). Consequently, we are unable to compare styles and
end up setting the 'effectiveAppearance' to 'appearance' (the computed value)
after r283858. Since the 'appearance' is not 'none', a native button is
briefly displayed during a hover animation, resulting in a white flash.

This was not an issue before r283858, as that change added logic to
always update the effectiveAppearance in RenderTheme::adjustStyle.
Before the change, the effectiveAppearance would simply be carried over
from the style before animation, and would remain 'none'.

To fix, only update the effectiveAppearance if its current value is
incorrect (auto, or styled). Additionally, use effectiveAppearance
more consistently, to ensure we do not incorrectly treat
style.appearance() as the used value.

Test: fast/forms/button-animation-appearance.html

* rendering/RenderTheme.cpp:
(WebCore::RenderTheme::adjustStyle):

LayoutTests:

Added a mismatch test to ensure that animating the background color
of a button drops native appearance. Verified that the test passes
before r283858, on Safari 15, and fails on r283858.

* fast/forms/button-animation-appearance-expected-mismatch.html: Added.
* fast/forms/button-animation-appearance.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284169 268f45cc-cd09-0410-ab3c-d52691b4dbfc