[webkitscmpy] Refactor PR branch management (Follow-up fix) https://bugs.webkit.org/show_bug.cgi?id=230432 <rdar://problem/83258413>
Reviewed by Stephanie Lewis.

When we are uploading a change that includes multiple commits,
the second commit will be on the same branch as the first, but that
branch will be editable. We should not declare that "no commits
have been found on an editable branch," because that's not true.

* Scripts/libraries/webkitscmpy/setup.py: Bump version.
* Scripts/libraries/webkitscmpy/webkitscmpy/__init__.py: Ditto.
* Scripts/libraries/webkitscmpy/webkitscmpy/program/branch.py:
(Branch.branch_point): Log the found commit for branches with multiple commits.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284263 268f45cc-cd09-0410-ab3c-d52691b4dbfc