[CMake] Align OptionsMac with Xcode build https://bugs.webkit.org/show_bug.cgi?id=231982
Reviewed by Don Olmstead.

.:

* Source/cmake/OptionsMac.cmake:
Ensure that the listed options (1) are not redundant with the defaults in WebKitFeatures.cmake and
(2) match the options for the Xcode build as closely as possible.

* Source/cmake/WebKitFeatures.cmake:
Remove an outdated feature, ENABLE_APPLE_PAY_SESSION_V9,
and let ENABLE_MAC_GESTURE_EVENTS be handled by PlatformEnable, since it depends on USE_APPLE_INTERNAL_SDK.

Source/WebCore:

* PlatformMac.cmake:
Add support for new files.

* platform/mediastream/cocoa/DisplayCaptureSourceCocoa.cpp:
Fix for r284295 -- ReplayKit is iOS and its headers shouldn't be included on Mac.

Source/WebKit:

* CMakeLists.txt:
Enable -fobjc-weak to support r284329.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284512 268f45cc-cd09-0410-ab3c-d52691b4dbfc