CSSPropertyZoom needs wrapper that ensures it's always blended into a positive value. https://bugs.webkit.org/show_bug.cgi?id=232020 <rdar://problem/84469930>
Reviewed by Antti Koivisto.

Source/WebCore:

Test: webanimations/zoom-animation-crash.html

The "zoom" CSS property is expected to be always larger than 0, so let's ensure we do not allow
values <= 0 while blending. To do so we repurpose NonNegativeFloatPropertyWrapper to now take
an argument specifying if the blended value should be non-negative or positive.

* animation/CSSPropertyAnimation.cpp:
(WebCore::FloatPropertyWrapper::FloatPropertyWrapper):
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):
(WebCore::NonNegativeFloatPropertyWrapper::NonNegativeFloatPropertyWrapper): Deleted.

LayoutTests:

Add a test that would crash prior to the source change.

* webanimations/zoom-animation-crash-expected.txt: Added.
* webanimations/zoom-animation-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284602 268f45cc-cd09-0410-ab3c-d52691b4dbfc