[WebXR] WebXR on Cocoa doesn't work with multisampled contexts https://bugs.webkit.org/show_bug.cgi?id=226687 <rdar://problem/78910868>
Reviewed by Tim Horton.

If the WebXR session requested an anti-aliased context, we'd
end up with an incomplete framebuffer because we were never
initialising the multisample buffers. This would assert on
debug builds and produce no output on release builds.

Fix this by initialising the multisample buffers, ensuring
we're targeting them in startFrame, and do the resolution
in endFrame.

* Modules/webxr/WebXROpaqueFramebuffer.cpp:
(WebCore::WebXROpaqueFramebuffer::startFrame):
(WebCore::WebXROpaqueFramebuffer::endFrame): While here, use NEAREST
sampling on the multisample resolution, the way we do in WebGL.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@284650 268f45cc-cd09-0410-ab3c-d52691b4dbfc